refactor: break the agent-lifecycle import cycle by extracting shared leaf modules (#2837)
agentLifecycle.js sat in two static import cycles — it imports both spawners
(agentCliSpawning.js, agentTuiSpawning.js) while both imported finalizeAgent /
releaseAgentLane back out of it — and papered over a third edge with
`await import('./agentManagement.js')` for handleOrphanedTask.

Move the pieces the lower layers need into leaf modules that nothing in the
cluster is imported by:

  agentFinalization.js          finalizeAgent, releaseAgentLane, the
                                success-criteria verdict and output-hook dispatch
  agentSummaryExtraction.js     extractFinalSummary, extractSimplifySummaries
  agentRunnerSync.js            syncRunnerAgents
  agentRunnerOutputBatchers.js  the per-agent runner output batchers

With those in place agentManagement.js no longer depends on agentLifecycle.js,
so three dynamic-import cycle dodges become plain static imports:
handleOrphanedTask, the spawner's init-time lifecycle/management lookups, and
flushRunnerOutputBatcher. cosAgentLifecycle.js keeps its dynamic delegation (it
is the state layer under agentManagement) but now targets agentManagement.js
directly instead of pulling the whole subAgentSpawner barrel.

agentLifecycle.js and subAgentSpawner.js re-export every moved symbol, so no
caller's import path changes. agentImportCycles.test.js re-derives the static
import graph from source and fails if a cycle is reintroduced.
