+++ .changelog/NEXT.md
@@ -55,3 +55,4 @@
 - **Commit the `server/` and `client/` lockfiles** — both were gitignored purely "to reduce merge noise" (`bbeab937d`), which left Dependabot able to see only the ~45 dependencies declared in their `package.json` files and blind to ~560 transitive packages; the repo's entire alert history is one alert against the root lockfile, whose tree has a single direct dep (pm2). CI's `npm install --prefix server|client` steps become `npm ci`, so the two trees holding all application code now resolve deterministically instead of re-resolving on every run and on every user's install. `setup-node`'s npm cache keys off all three lockfiles. `browser/package-lock.json` stays ignored (its `node_modules` is committed instead). Unrelated to the npm-workspaces revert in `b743222be`, which was about exFAT's lack of symlink support.
 - **Apply the mergeable Dependabot bumps in one commit** — server `vitest`/`@vitest/coverage-v8` 4.1.9→4.1.10 and `node-pty` 1.2.0-beta.10→beta.14; client `react`/`react-dom` 19.2.7→19.2.8, `vite` 8.0.16→8.1.5, and `@tailwindcss/postcss` 4.3.1→4.3.3. Applied together with regenerated lockfiles because the individual PRs predate lockfile tracking and each ERESOLVE'd in isolation: the two vitest packages must move in lockstep, `@tailwindcss/postcss` was stranded a patch behind the already-merged `tailwindcss` 4.3.3, and Dependabot's `react-dom` PR would have left `react` mismatched at 19.2.7. Holding back `undici` 8 (breaks the DNS-rebinding connect-time IP pinning in `safeUrlFetch.js`, #1859) and `eslint`/`@eslint/js` 10 (`eslint-plugin-react@7.37.5` peers at `^9.7`).
 - **[issue-2845] Standardize the `loraTraining` sample 404 error envelope** — new `sendErrorResponse`/`buildErrorEnvelope` helpers in `server/lib/errorHandler.js` build the shared `{ error, code, timestamp }` body in one place; `asyncHandler`, `errorMiddleware`, the `res.sendFile` 404 callback in `server/routes/loraTraining.js`, and the `authGate` 401/403 short-circuits now all route through it, so responses raised outside a handler's catch no longer omit `code`/`timestamp`. Mid-response failures (SSE/streaming routes) now delegate to Express's finalhandler so the socket is destroyed instead of the request hanging.
+- **[issue-2837] Break the agent-lifecycle circular-dependency cluster** — `finalizeAgent`/`releaseAgentLane` (plus the success-criteria + output-hook helpers) move to `server/services/agentFinalization.js`, the summary extractors to `agentSummaryExtraction.js`, runner recovery to `agentRunnerSync.js`, and the runner output batchers to `agentRunnerOutputBatchers.js`. With those leaves in place the two static cycles (`agentLifecycle` ↔ `agentCliSpawning`, `agentLifecycle` ↔ `agentTuiSpawning`) are gone, and three `await import()` cycle-dodges — `handleOrphanedTask`, the spawner's init-time lifecycle/management lookups, and `flushRunnerOutputBatcher` — become plain static imports. `agentLifecycle.js` still re-exports every moved symbol, so no import path changes for callers. New `agentImportCycles.test.js` re-derives the static import graph and fails if a cycle is reintroduced.
