+++ server/services/agentRunnerSync.js
@@ -1,0 +1,72 @@
+/**
+ * Runner agent recovery.
+ *
+ * Rehydrates the in-memory `runnerAgents` map from the CoS Runner after a
+ * server restart, so completion events for agents spawned before the restart
+ * still land. Extracted from `agentLifecycle.js` (issue #2837): both
+ * `agentManagement.js` and `subAgentSpawner.js` need it, and importing it from
+ * the lifecycle orchestrator dragged that whole module graph into a cycle.
+ *
+ * Leaf with respect to the agent cluster — must not import `agentLifecycle.js`
+ * or `agentManagement.js` (enforced by `agentImportCycles.test.js`).
+ */
+
+import { getActiveAgentsFromRunner } from './cosRunnerClient.js';
+import { isInternalTaskId } from '../lib/taskParser.js';
+import { runnerAgents } from './agentState.js';
+
+/**
+ * Sync running agents from the runner (recovery after server restart).
+ * This allows us to receive completion events for agents spawned before restart.
+ */
+export async function syncRunnerAgents() {
+  const agents = await getActiveAgentsFromRunner().catch(err => {
+    console.error(`❌ Failed to get active agents from runner: ${err.message}`);
+    return [];
+  });
+  if (agents.length === 0) return 0;
+
+  console.log(`🔄 Syncing ${agents.length} running agents from CoS Runner`);
+
+  // Get all tasks to find task data for each agent
+  const { getAllTasks } = await import('./cos.js');
+  const allTasksData = await getAllTasks().catch(() => ({ user: {}, cos: {} }));
+
+  // Build a task lookup map from all task sources, tagging each with its taskType
+  const taskMap = new Map();
+  const addTasks = (groupedTasks, taskType) => {
+    if (!groupedTasks) return;
+    for (const tasks of Object.values(groupedTasks)) {
+      if (Array.isArray(tasks)) {
+        for (const task of tasks) {
+          taskMap.set(task.id, { ...task, taskType });
+        }
+      }
+    }
+  };
+
+  addTasks(allTasksData.user?.grouped, 'user');
+  addTasks(allTasksData.cos?.grouped, 'internal');
+
+  let syncedCount = 0;
+  for (const agent of agents) {
+    // Only sync if not already tracked
+    if (!runnerAgents.has(agent.id)) {
+      const task = taskMap.get(agent.taskId);
+
+      const inferredType = isInternalTaskId(agent.taskId) ? 'internal' : 'user';
+      runnerAgents.set(agent.id, {
+        taskId: agent.taskId,
+        task: task || { id: agent.taskId, taskType: inferredType, description: 'Recovered from runner' },
+        runId: null, // Run tracking may be lost on restart
+        model: null,
+        hasStartedWorking: true,
+        startedAt: agent.startedAt
+      });
+      console.log(`🔄 Recovered agent ${agent.id} (task: ${agent.taskId})`);
+      syncedCount++;
+    }
+  }
+
+  return syncedCount;
+}
